requirements.txt
asgiref==3.2.6
beautifulsoup4==4.8.2
Django==3.0.4
django-bootstrap4==1.1.1
django-crispy-forms==1.9.0
django-datatables-view==1.19.1
django-jquery==3.1.0
django-registration-redux==2.7
django-scheduler==0.9.3
django-taggit==1.2.0
django-tagulous==0.14.1
django-tempus-dominus==5.1.2.13
dnspython==1.16.0
icalendar==4.0.5
mysql-connector-python==8.0.19
postgres==3.0.0
protobuf==3.6.1
psycopg2-binary==2.8.4
psycopg2-pool==1.1
python-dateutil==2.8.1
pytz==2019.3
simplejson==3.17.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
